deactive account message to tg

+++ src/modules/cron/cron.service.ts
@@ -5,13 +5,15 @@ import { PrismaService } from '@/src/core/prisma/prisma.service'
 
 import { MailService } from '../libs/mail/mail.service'
 import { StorageService } from '../libs/storage/storage.service'
+import { TelegramService } from '../libs/telegram/telegram.service'
 
 @Injectable()
 export class CronService {
 	public constructor(
 		private readonly prismaService: PrismaService,
 		private readonly mailService: MailService,
-		private readonly storageService: StorageService
+		private readonly storageService: StorageService,
+		private readonly telegramService: TelegramService
 	) {}
 
 	@Cron('0 0 * * *')
@@ -26,6 +28,10 @@ export class CronService {
 				deactivatedAt: {
 					lte: sevenDaysAgo
 				}
+			},
+			include: {
+				notificationSettings: true,
+				stream: true
 			}
 		})
 
@@ -41,7 +47,20 @@ export class CronService {
 		for (const user of deactivatedAccounts) {
 			await this.mailService.sendAccountDeletion(user.email)
 
-			this.storageService.remove(user.avatar)
+			if (
+				user.notificationSettings.telegramNotifications &&
+				user.telegramId
+			) {
+				await this.telegramService.sendAccountDeletion(user.telegramId)
+			}
+
+			if (user.avatar) {
+				this.storageService.remove(user.avatar)
+			}
+
+			if (user.stream.thumbnailUrl) {
+				this.storageService.remove(user.stream.thumbnailUrl)
+			}
 		}
 	}
 }
